Extracted the register page.
